bug: initial IP assignment fails adding dns record to google cloud dns api
Upon assignment of a public IP for the first time to a new service
there will be no previous public IP set on the service. In this case we
were trying to send a google cloud DNS update changeset that included
both an "addition" for the newly assigned public IP and a "deletion"
that includes a nil list of public IPs. This was causing google cloud
dns api to fail the entire changeset with an error such as:

```
2016/11/29 01:04:17 Got Possible Service update prometheus-delete.development.gce.panth.io.
2016/11/29 01:04:17 Error creating change: googleapi: Error 400: Invalid
value for 'entity.change.deletions[0].rrdata': '<empty>', invalid
```